+++ dev/bench/bench_nms.py
@@ -1,7 +1,7 @@
 """
 Timed best=408.694 µs, mean=412.551 ± 3.9 µs for impl=cython_cpu,type=ndarray
-Timed best=599.641 µs, mean=943.981 ± 418.9 µs for impl=cython_cpu,type=tensor0
-Timed best=2.426 ms, mean=2.539 ± 0.1 ms for impl=torch,type=tensor0
+Timed best=599.641 µs, mean=943.981 ± 418.9 µs for impl=cython_cpu,type=tensor-cuda0
+Timed best=2.426 ms, mean=2.539 ± 0.1 ms for impl=torch,type=tensor-cuda0
 Timed best=4.455 ms, mean=4.686 ± 0.2 ms for impl=torch,type=ndarray
 Timed best=12.725 ms, mean=13.007 ± 0.2 ms for impl=numpy,type=ndarray
 """
@@ -24,6 +24,21 @@ def ensure_numpy_indices(keep):
     #     keep = keep
     keep = np.array(sorted(np.array(keep).ravel()))
     return keep
+
+
+def make_idstr(d):
+    """
+    Make full-length-key id-string
+    """
+    if d is None:
+        return ''
+    elif isinstance(d, str):
+        return d
+    elif len(d) == 0:
+        return ''
+    if not isinstance(d, ub.odict):
+        d = ub.odict(sorted(d.items()))
+    return ub.urepr(d, itemsep='', nobr=True, explicit=True, nl=0, si=True)
 
 
 def benchamrk_det_nms():
@@ -90,14 +105,14 @@ def benchamrk_det_nms():
     basis = {
         'type': [
             'ndarray',
-            'tensor',
-            'tensor0'
+            'tensor-cpu',
+            'tensor-cuda0'
         ],
         # 'daq': [True, False],
         # 'daq': [False],
         # 'device': [None],
-        # 'impl': valid_impls,
-        'impl': valid_impls + ['auto'],
+        'impl': valid_impls,
+        # 'impl': valid_impls + ['auto'],
     }
 
     if ub.argflag('--daq'):
@@ -130,15 +145,15 @@ def benchamrk_det_nms():
         REMOVE_SLOW = True
         if REMOVE_SLOW:
             known_bad = [
-                {'impl': 'torch', 'type': 'tensor'},
-                {'impl': 'numpy', 'type': 'tensor'},
-                # {'impl': 'cython_gpu', 'type': 'tensor'},
-                {'impl': 'cython_cpu', 'type': 'tensor'},
+                {'impl': 'torch', 'type': 'tensor-cpu'},
+                {'impl': 'numpy', 'type': 'tensor-cpu'},
+                # {'impl': 'cython_gpu', 'type': 'tensor-cpu'},
+                {'impl': 'cython_cpu', 'type': 'tensor-cpu'},
 
-                # {'impl': 'torch', 'type': 'tensor0'},
-                {'impl': 'numpy', 'type': 'tensor0'},
-                # {'impl': 'cython_gpu', 'type': 'tensor0'},
-                # {'impl': 'cython_cpu', 'type': 'tensor0'},
+                # {'impl': 'torch', 'type': 'tensor-cuda0'},
+                {'impl': 'numpy', 'type': 'tensor-cuda0'},
+                # {'impl': 'cython_gpu', 'type': 'tensor-cuda0'},
+                # {'impl': 'cython_cpu', 'type': 'tensor-cuda0'},
 
                 {'impl': 'torchvision', 'type': 'ndarray'},
             ]
@@ -196,11 +211,10 @@ def benchamrk_det_nms():
         typed_data = {}
         # ----------------------------------
 
-        import netharn as nh
         for combo in combos:
             print('combo = {}'.format(ub.urepr(combo, nl=0)))
 
-            label = nh.util.make_idstr(combo)
+            label = make_idstr(combo)
             mode = combo.copy()
 
             # if mode['impl'] == 'cython_gpu':
@@ -213,10 +227,10 @@ def benchamrk_det_nms():
             else:
                 if mode_type == 'ndarray':
                     dets = np_dets.numpy()
-                elif mode_type == 'tensor':
-                    dets = np_dets.tensor(None)
-                elif mode_type == 'tensor0':
-                    dets = np_dets.tensor(0)
+                elif mode_type == 'tensor-cpu':
+                    dets = np_dets.tensor(device='cpu')
+                elif mode_type == 'tensor-cuda0':
+                    dets = np_dets.tensor(device=0)
                 else:
                     raise KeyError
                 typed_data[mode_type] = dets
@@ -294,7 +308,7 @@ def benchamrk_det_nms():
             for key, vals in ydata.items():
                 score_wrt_x[key] = vals[pos]
 
-            typekeys = ['tensor0', 'tensor', 'ndarray']
+            typekeys = ['tensor-cuda0', 'tensor-cpu', 'ndarray']
             type_groups = dict([
                 (b, ub.group_items(score_wrt_x, lambda y: y.endswith(b))[True])
                 for b in typekeys
@@ -322,7 +336,7 @@ def benchamrk_det_nms():
                     ordered_impls.append(d['impl'])
 
                 ordered_impls = list(ub.oset(ordered_impls) - {'auto'})
-                ordered_impls2.pop('auto')
+                ordered_impls2.pop('auto', None)
                 record('        # {}'.format(ub.urepr(ordered_impls2, precision=1, nl=0, explicit=True)))
                 record('        preference = {}'.format(ub.urepr(ordered_impls, nl=0)))
     record('### end times of interest ')