+++ kwimage/transform.py
@@ -444,7 +444,7 @@ class Projective(Linear):
         >>> print('homog = {}'.format(ub.urepr(homog.tolist(), nl=1)))
         >>> # This could be prettier
         >>> texts = {
-        >>>     'Translation': sympy.pretty(R, use_unicode=0),
+        >>>     'Translation': sympy.pretty(T, use_unicode=0),
         >>>     'Rotation': sympy.pretty(R, use_unicode=0),
         >>>     'shEar-X': sympy.pretty(E, use_unicode=0),
         >>>     'Scale': sympy.pretty(S, use_unicode=0),
@@ -1090,10 +1090,10 @@ class Affine(Projective):
         >>> print('aff = {}'.format(ub.urepr(aff.tolist(), nl=1)))
         >>> # This could be prettier
         >>> texts = {
-        >>>     'Translation': sympy.pretty(R),
-        >>>     'Rotation': sympy.pretty(R),
-        >>>     'shEar-X': sympy.pretty(E),
-        >>>     'Scale': sympy.pretty(S),
+        >>>     'Translation': sympy.pretty(sympy.Matrix(T), use_unicode=0),
+        >>>     'Rotation': sympy.pretty(sympy.Matrix(R), use_unicode=0),
+        >>>     'shEar-X': sympy.pretty(sympy.Matrix(E), use_unicode=0),
+        >>>     'Scale': sympy.pretty(sympy.Matrix(S), use_unicode=0),
         >>> }
         >>> print(ub.urepr(texts, nl=2, sv=1))
         >>> equation_stack = []
@@ -1103,7 +1103,7 @@ class Affine(Projective):
         >>>     render_canvas = kwimage.imresize(render_canvas, scale=1.3)
         >>>     equation_stack.append(render_canvas)
         >>> equation_canvas = kwimage.stack_images(equation_stack, pad=10, axis=1, bg_value='kitware_gray')
-        >>> render_canvas = kwimage.draw_text_on_image(None, sympy.pretty(aff), color='kw_blue', fontScale=1.0)
+        >>> render_canvas = kwimage.draw_text_on_image(None, sympy.pretty(sympy.Array(aff), use_unicode=0), color='kw_blue', fontScale=1.0)
         >>> render_canvas = kwimage.draw_header_text(render_canvas, 'Full Equation With Pre-Shift', color='kw_green')
         >>> # xdoctest: -REQUIRES(module:sympy)
         >>> # xdoctest: +REQUIRES(--show)
@@ -1555,7 +1555,6 @@ class Affine(Projective):
             if ub.iterable(kw['scale']) and (not isinstance(kw['scale'], tuple) and len(kw['scale']) == 2):
                 raise NotImplementedError
             else:
-                print(kw['scale'])
                 xscale_dist = _coerce_distri(kw['scale'])
                 yscale_dist = xscale_dist
         else:
@@ -2008,52 +2007,100 @@ class Affine(Projective):
             cv2_method = inlier_method_lut[inlier_method]
             A, mask = cv2.estimateAffine2D(pts1, pts2, method=cv2_method)
 
-        x1_mn = pts1[:, 0]
-        y1_mn = pts1[:, 1]
-        x2_mn = pts2[:, 0]
-        y2_mn = pts2[:, 1]
-        num_pts = x1_mn.shape[0]
-        Mx6 = np.empty((2 * num_pts, 6), dtype=float)
-        b = np.empty((2 * num_pts, 1), dtype=float)
-        for ix in range(num_pts):  # Loop over inliers
-            # Concatenate all 2x9 matrices into an Mx6 matrix
-            x1 = x1_mn[ix]
-            x2 = x2_mn[ix]
-            y1 = y1_mn[ix]
-            y2 = y2_mn[ix]
-            Mx6[ix * 2]     = (x1, y1, 0, 0, 1, 0)
-            Mx6[ix * 2 + 1] = ( 0, 0, x1, y1, 0, 1)
-            b[ix * 2] = x2
-            b[ix * 2 + 1] = y2
+        num_pts = len(pts1)
+        if num_pts == 0:
+            # Regularize degenerate 0 correspondence case
+            mat = np.eye(3)  # Identity if no points
+        elif num_pts == 1:
+            # Regularize degenerate 1 correspondence case
+            # Only translation possible
+            t1 = np.mean(pts1, axis=0)
+            t2 = np.mean(pts2, axis=0)
+            tx, ty = t2 - t1
+            mat = np.array([
+                [1, 0, tx],
+                [0, 1, ty],
+                [0, 0, 1]
+            ])
+        elif num_pts == 2:
+            # Regularize degenerate 2 correspondence case
+            # For 2 points, compute rotation and scaling
+            # Center points to estimate rotation/scale
+            t1 = np.mean(pts1, axis=0)
+            t2 = np.mean(pts2, axis=0)
+            centered1 = pts1 - t1
+            centered2 = pts2 - t2
 
-        M = Mx6
-        try:
-            USVt = np.linalg.svd(M, full_matrices=True, compute_uv=True)
-        except MemoryError:
-            import scipy.sparse as sps
-            import scipy.sparse.linalg as spsl
-            M_sparse = sps.lil_matrix(M)
-            USVt = spsl.svds(M_sparse)
-        except np.linalg.LinAlgError:
-            raise
-        except Exception:
-            raise
+            # Solve for rotation and scaling using Kabsch algorithm (simplified)
+            # https://en.wikipedia.org/wiki/Kabsch_algorithm
+            H = centered1.T @ centered2  # Cross-covariance matrix
+            U, _, Vt = np.linalg.svd(H)
+            R = Vt.T @ U.T  # Rotation matrix
 
-        U, s, Vt = USVt
+            # Handle reflection case (enforce proper rotation)
+            if np.linalg.det(R) < 0:
+                Vt[-1, :] *= -1
+                R = Vt.T @ U.T
 
-        # Inefficient, but the math works
-        # We want to solve Ax=b (where A is the Mx6 in this case)
-        # Ax = b
-        # (U S V.T) x = b
-        # x = (U.T inv(S) V) b
-        Sinv = np.zeros((len(Vt), len(U)))
-        Sinv[np.diag_indices(len(s))] = 1 / s
-        a = Vt.T.dot(Sinv).dot(U.T).dot(b).T[0]
-        mat = np.array([
-            [a[0], a[1], a[4]],
-            [a[2], a[3], a[5]],
-            [   0, 0, 1],
-        ])
+            # Estimate scaling
+            s = np.trace(H @ R) / np.trace(centered1.T @ centered1)
+
+            # Build full transform: [sR | t2 - sR t1]
+            mat = np.eye(3)
+            mat[:2, :2] = s * R
+            mat[:2, 2] = t2 - (s * R) @ t1
+        else:
+            x1_mn = pts1[:, 0]
+            y1_mn = pts1[:, 1]
+            x2_mn = pts2[:, 0]
+            y2_mn = pts2[:, 1]
+            num_pts = x1_mn.shape[0]
+            Mx6 = np.empty((2 * num_pts, 6), dtype=float)
+            b = np.empty((2 * num_pts, 1), dtype=float)
+            for ix in range(num_pts):  # Loop over inliers
+                # Concatenate all 2x9 matrices into an Mx6 matrix
+                x1 = x1_mn[ix]
+                x2 = x2_mn[ix]
+                y1 = y1_mn[ix]
+                y2 = y2_mn[ix]
+                Mx6[ix * 2]     = (x1, y1, 0, 0, 1, 0)
+                Mx6[ix * 2 + 1] = ( 0, 0, x1, y1, 0, 1)
+                b[ix * 2] = x2
+                b[ix * 2 + 1] = y2
+
+            M = Mx6
+            try:
+                USVt = np.linalg.svd(M, full_matrices=True, compute_uv=True)
+            except MemoryError:
+                import scipy.sparse as sps
+                import scipy.sparse.linalg as spsl
+                M_sparse = sps.lil_matrix(M)
+                USVt = spsl.svds(M_sparse)
+            except np.linalg.LinAlgError:
+                raise
+            except Exception:
+                raise
+
+            U, s, Vt = USVt
+
+            if np.isclose(s[-1], 0):
+                # SVD solution failed, fallback on lstsq
+                a, _, _, _ = np.linalg.lstsq(M, b, rcond=None)
+                a = a.ravel()
+            else:
+                # Inefficient, but the math works
+                # We want to solve Ax=b (where A is the Mx6 in this case)
+                # Ax = b
+                # (U S V.T) x = b
+                # x = (U.T inv(S) V) b
+                Sinv = np.zeros((len(Vt), len(U)))
+                Sinv[np.diag_indices(len(s))] = 1 / s
+                a = Vt.T.dot(Sinv).dot(U.T).dot(b).T[0]
+            mat = np.array([
+                [a[0], a[1], a[4]],
+                [a[2], a[3], a[5]],
+                [   0, 0, 1],
+            ])
         return Affine(mat)
 
     @classmethod